minor cleanup

+++ Validation/GlobalHits/test/valid_global.csh
@@ -4,9 +4,9 @@ eval `scramv1 ru -csh`
 
 echo "============> Validating Global Sim Hits <============"
 echo "......producing output file with this release"
-rm MessageLogger.log
-rm output.log
-rm FrameworkJobReport.xml
+if ( -e MessageLogger.log ) rm MessageLogger.log
+if ( -e output.log ) rm output.log
+if ( -e FrameworkJobReport.xml ) rm FrameworkJobReport.xml
 cmsRun -p EvtGen+DetSim+Global.cfg >& output.log
 echo "......comparing against reference file from previous release"
 root -b -q MakeValidation.C
